# Cache common JsValue allocations