--- a/Jint/Engine.cs
+++ b/Jint/Engine.cs
@@ -45,22 +45,22 @@ namespace Jint
 
         internal static Dictionary<Type, Func<Engine, object, JsValue>> TypeMappers = new Dictionary<Type, Func<Engine, object, JsValue>>()
         {
-            { typeof(bool), (Engine engine, object v) => new JsValue((bool)v) },
-            { typeof(byte), (Engine engine, object v) => new JsValue((byte)v) },
-            { typeof(char), (Engine engine, object v) => new JsValue((char)v) },
+            { typeof(bool), (Engine engine, object v) => (bool) v ? JsValue.True : JsValue.False },
+            { typeof(byte), (Engine engine, object v) => JsValue.FromInt((byte)v) },
+            { typeof(char), (Engine engine, object v) => JsValue.FromChar((char)v) },
             { typeof(DateTime), (Engine engine, object v) => engine.Date.Construct((DateTime)v) },
             { typeof(DateTimeOffset), (Engine engine, object v) => engine.Date.Construct((DateTimeOffset)v) },
-            { typeof(decimal), (Engine engine, object v) => new JsValue((double)(decimal)v) },
-            { typeof(double), (Engine engine, object v) => new JsValue((double)v) },
-            { typeof(Int16), (Engine engine, object v) => new JsValue((Int16)v) },
-            { typeof(Int32), (Engine engine, object v) => new JsValue((Int32)v) },
-            { typeof(Int64), (Engine engine, object v) => new JsValue((Int64)v) },
-            { typeof(SByte), (Engine engine, object v) => new JsValue((SByte)v) },
-            { typeof(Single), (Engine engine, object v) => new JsValue((Single)v) },
-            { typeof(string), (Engine engine, object v) => new JsValue((string)v) },
-            { typeof(UInt16), (Engine engine, object v) => new JsValue((UInt16)v) },
-            { typeof(UInt32), (Engine engine, object v) => new JsValue((UInt32)v) },
-            { typeof(UInt64), (Engine engine, object v) => new JsValue((UInt64)v) },
+            { typeof(decimal), (Engine engine, object v) => (JsValue) (double)(decimal)v },
+            { typeof(double), (Engine engine, object v) => (JsValue)(double)v },
+            { typeof(Int16), (Engine engine, object v) => JsValue.FromInt((Int16)v) },
+            { typeof(Int32), (Engine engine, object v) => JsValue.FromInt((Int32)v) },
+            { typeof(Int64), (Engine engine, object v) => (JsValue)(Int64)v },
+            { typeof(SByte), (Engine engine, object v) => JsValue.FromInt((SByte)v) },
+            { typeof(Single), (Engine engine, object v) => (JsValue)(Single)v },
+            { typeof(string), (Engine engine, object v) => (JsValue) (string)v },
+            { typeof(UInt16), (Engine engine, object v) => JsValue.FromInt((UInt16)v) },
+            { typeof(UInt32), (Engine engine, object v) => JsValue.FromInt((UInt32)v) },
+            { typeof(UInt64), (Engine engine, object v) => JsValue.FromInt((UInt64)v) },
             { typeof(JsValue), (Engine engine, object v) => (JsValue)v },
             { typeof(System.Text.RegularExpressions.Regex), (Engine engine, object v) => engine.RegExp.Construct((System.Text.RegularExpressions.Regex)v, "") }
         };
@@ -257,9 +257,14 @@ namespace Jint
             return SetValue(name, new JsValue(value));
         }
 
+        public Engine SetValue(string name, int value)
+        {
+            return SetValue(name, JsValue.FromInt(value));
+        }
+
         public Engine SetValue(string name, bool value)
         {
-            return SetValue(name, new JsValue(value));
+            return SetValue(name, value ? JsValue.True : JsValue.False);
         }
 
         public Engine SetValue(string name, JsValue value)
@@ -539,7 +544,7 @@ namespace Jint
                 {
                     return baseValue;
                 }
-                
+
                 if (reference.HasPrimitiveBase() == false)
                 {
                     var o = TypeConverter.ToObject(this, baseValue);