Here we get a really nice reduction of allocations and a speed boost by catching all sorts of implicit JsValue allocations. I generally don't like implicit operators but it was quite easy to redirect them to use caching. 

As JsValue is immutable and attached to certain object, we can easily pair them lazily and reuse. Also caching common int, double and char maps.

## UncacheableExpressionsBenchmark

### Before 

|    Method |   N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |--------:|---------:|---------:|------------:|-----------:|----------:|
| Benchmark | 500 | 1.214 s | 0.0755 s | 0.1083 s | 202687.5000 | 55312.5000 |  903.2 MB |

### After

|    Method |   N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |--------:|---------:|---------:|------------:|-----------:|----------:|
| Benchmark | 500 | 1.131 s | 0.0086 s | 0.0123 s | 189937.5000 | 54187.5000 |  817.6 MB |