@lahma 2017-12-24
@sebastienros does this look mergeable? I've kept additions internal so shouldn't cause too much confusion.