@sebastienros 2017-12-24
I'd like to keep this change for later. i might refactor JsValue to be a base class for ObjectInstance too making everything a JsValue.