@lahma 2017-12-25
Could you perhaps refactor after merge? The API does not actually change as the cache helpers are internal. When refactoring can then easily reflect against the "ideal" numbers.

With all these PRs the problem is not having previous steps merged causing a lot of noise in profiling. This JsValue caching does a big difference with 10% allocation reduction.