@lahma 2018-01-02
As I found out that Literal.Cacheable is never set to true, the optimization path never kicks in. I've changed cacheable to ignored as the values are literals. Even better performance now:

|    Method |   N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |--------:|---------:|---------:|------------:|-----------:|----------:|
| Benchmark | 500 | 1.058 s | 0.0096 s | 0.0140 s | 187437.5000 | 62437.5000 |  812.9 MB |